+++ sickle/models.py
@@ -1,15 +1,30 @@
 # coding: utf-8
 """
-    models
-    ~~~~~~
+    sickle.models
+    ~~~~~~~~~~~~~
 
     Collects classes for OAI-specific entities.
 
-    :copyright: Copright 2013 Mathias Loesch
+    :copyright: Copyright 2015 Mathias Loesch
 """
 
 from lxml import etree
+
 from .utils import get_namespace, xml_to_dict
+
+
+class ResumptionToken(object):
+    """Represents a resumption token."""
+
+    def __init__(self, token='', cursor='', complete_list_size='',
+                 expiration_date=''):
+        self.token = token
+        self.cursor = cursor
+        self.complete_list_size = complete_list_size
+        self.expiration_date = expiration_date
+
+    def __repr__(self):
+        return '<ResumptionToken %s>' % self.token
 
 
 class OAIItem(object):
@@ -19,6 +34,7 @@ class OAIItem(object):
     :param strip_ns: Flag for whether to remove the namespaces from the
                      element names in the dictionary representation.
     """
+
     def __init__(self, xml, strip_ns=True):
         super(OAIItem, self).__init__()
 
@@ -43,11 +59,12 @@ class Identify(OAIItem):
     """Represents an Identify container.
 
     This object differs from the other entities in that is has to be created
-    from a :class:`sickle.app.OAIResponse` instead of an XML element.
+    from a :class:`sickle.response.OAIResponse` instead of an XML element.
 
     :param identify_response: The response for an Identify request.
-    :type identify_response: :class:`sickle.app.OAIResponse`
+    :type identify_response: :class:`sickle.OAIResponse`
     """
+
     def __init__(self, identify_response):
         super(Identify, self).__init__(identify_response.xml, strip_ns=True)
         self.xml = self.xml.find('.//' + self._oai_namespace + 'Identify')
@@ -66,7 +83,9 @@ class Header(OAIItem):
     """Represents an OAI Header.
 
     :param header_element: The XML element 'header'.
+    :type header_element: :class:`lxml.etree._Element
     """
+
     def __init__(self, header_element):
         super(Header, self).__init__(header_element, strip_ns=True)
         self.deleted = self.xml.attrib.get('status') == 'deleted'
@@ -95,9 +114,11 @@ class Record(OAIItem):
     """Represents an OAI record.
 
     :param record_element: The XML element 'record'.
+    :type record_element: :class:`lxml.etree._Element`
     :param strip_ns: Flag for whether to remove the namespaces from the
                      element names.
     """
+
     def __init__(self, record_element, strip_ns=True):
         super(Record, self).__init__(record_element, strip_ns=strip_ns)
         self.header = Header(self.xml.find(
@@ -126,7 +147,9 @@ class Set(OAIItem):
     """Represents an OAI set.
 
     :param set_element: The XML element 'set'.
+    :type set_element: :class:`lxml.etree._Element
     """
+
     def __init__(self, set_element):
         super(Set, self).__init__(set_element, strip_ns=True)
         self._set_dict = xml_to_dict(self.xml, strip_ns=True)
@@ -134,7 +157,7 @@ class Set(OAIItem):
             setattr(self, k.replace('-', '_'), v[0])
 
     def __repr__(self):
-        return u'<Set %s>'.encode("utf8") % self.setName
+        return u'<Set %s>'.encode('utf8') % self.setName
 
     def __iter__(self):
         return self._set_dict.iteritems()
@@ -144,7 +167,9 @@ class MetadataFormat(OAIItem):
     """Represents an OAI MetadataFormat.
 
     :param mdf_element: The XML element 'metadataFormat'.
+    :type mdf_element: :class:`lxml.etree._Element
     """
+
     def __init__(self, mdf_element):
         super(MetadataFormat, self).__init__(mdf_element, strip_ns=True)
         #: The prefix of this format.
@@ -153,7 +178,7 @@ class MetadataFormat(OAIItem):
             setattr(self, k.replace('-', '_'), v[0])
 
     def __repr__(self):
-        return u'<MetadataFormat %s>'.encode("utf8") % self.metadataPrefix
+        return u'<MetadataFormat %s>'.encode('utf8') % self.metadataPrefix
 
     def __iter__(self):
         return self._mdf_dict.iteritems()