+++ CHANGES.rst
@@ -1,10 +1,22 @@
 Changelog
 =========
 
+
+Version 0.4
+-----------
+
+May 31, 2015
+
+- bug fix: resumptionToken parameter is exclusive
+- added support for harvesting complete OAI-XML responses
+
+
 Version 0.3
 -----------
 
-- added support for protected OAI interfaces (basic auth)
+April 17, 2013
+
+- added support for protected OAI interfaces (basic authentication)
 - made class mapping for OAI elements configurable
 - added options for HTTP timeout and max retries
 - added handling of HTTP 503 responses
@@ -13,9 +25,12 @@ Version 0.3
 Version 0.2
 -----------
 
+February 26, 2013
+
 - OAI items are now represented as their own classes instead of XML elements
 - library raises OAI-specific exceptions
 - made lxml a required dependency
+
 
 Version 0.1
 -----------